emotion/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
#tensorflow
transformers
protobuf==4.21
